pyproject.toml
[tool.poetry]
name = "mnist"
version = "1.0.0"
description = "MNIST recognition"
authors = ["BiscuitsLayer <vinogradov.so@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.5.0"
fire = "^0.5.0"
dvc = "^3.36.1"
dvc-gdrive = "^3.0.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
torch = "^2.1.0"
torchvision = "^0.16.0"
matplotlib = "^3.8.0"

[tool.poetry.scripts]
mnist_train = 'mnist.train:main'
mnist_infer = 'mnist.infer:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
